Feature/add support for ical (#397)
* Implement support for ics-attachments

* Implement tests for attachments and ical validity

* Make changes according to code review
